video/requirements.txt
pexels-api>=1.0.0
moviepy>=1.0.3
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0 